refactor: lift audio playback into Mithril.Shared
Promotes AlarmSoundPlayer + IPlaybackHandle from Samwise.Module to a
generic Mithril.Shared.Audio.AudioPlayer, and consolidates the duplicated
FlashWindowEx P/Invoke into Mithril.Shared.Wpf.WindowFlasher.

- Closes the shell-imports-Samwise leak (Program.cs no longer carries
  `using Samwise.Alarms`).
- Drops Gandalf.Module's project reference to Samwise.Module — Gandalf
  was only depending on Samwise to reach the audio primitive.
- NAudio package reference moves from two module projects to one
  shared project; Samwise/Gandalf no longer reference NAudio directly.
- Pure refactor: no behavior change. AlarmService (Samwise) and
  TimerAlarmService (Gandalf) keep their domain coupling and are
  unchanged apart from the using swap.
