+++ src/Gandalf.Module/Views/GandalfSettingsView.xaml.cs
@@ -1,9 +1,9 @@
 using System.Windows;
 using Gandalf.Domain;
 using Gandalf.ViewModels;
+using Mithril.Shared.Audio;
 using Mithril.Shared.Settings;
 using Microsoft.Win32;
-using Samwise.Alarms;
 
 namespace Gandalf.Views;
 
@@ -24,7 +24,7 @@ public partial class GandalfSettingsView : System.Windows.Controls.UserControl
         var dlg = new OpenFileDialog
         {
             Title = "Choose an alarm sound",
-            Filter = AlarmSoundPlayer.OpenFileFilter,
+            Filter = AudioPlayer.OpenFileFilter,
         };
         if (dlg.ShowDialog() == true) settings.SoundFilePath = dlg.FileName;
     }
@@ -34,7 +34,7 @@ public partial class GandalfSettingsView : System.Windows.Controls.UserControl
         var s = CurrentSettings;
         if (s is null) return;
         _testHandle?.Stop();
-        _testHandle = AlarmSoundPlayer.Play(s.SoundFilePath, (float)s.AlarmVolume, "gandalf");
+        _testHandle = AudioPlayer.Play(s.SoundFilePath, (float)s.AlarmVolume, "gandalf");
     }
 
     private void ClearSound_Click(object sender, RoutedEventArgs e)
