+++ src/Mithril.Shared/Audio/IPlaybackHandle.cs
@@ -1,0 +1,7 @@
+namespace Mithril.Shared.Audio;
+
+public interface IPlaybackHandle : IDisposable
+{
+    void Stop();
+    bool IsPlaying { get; }
+}
