+++ src/Samwise.Module/Alarms/AlarmService.cs
@@ -1,13 +1,14 @@
-using System.Runtime.InteropServices;
 using System.Windows;
 using System.Windows.Threading;
+using Mithril.Shared.Audio;
+using Mithril.Shared.Wpf;
 using Samwise.State;
 
 namespace Samwise.Alarms;
 
 public sealed record ActiveAlarm(string Key, string CharName, string CropType, DateTimeOffset Triggered);
 
-public sealed partial class AlarmService : IDisposable
+public sealed class AlarmService : IDisposable
 {
     private readonly GardenStateMachine _state;
     private readonly SamwiseSettings _settings;
@@ -77,13 +78,13 @@ public sealed partial class AlarmService : IDisposable
         Dispatch(() =>
         {
             StopPlayback(alarm.Key);
-            var handle = AlarmSoundPlayer.Play(soundFilePath, (float)_settings.Alarms.AlarmVolume, "samwise");
+            var handle = AudioPlayer.Play(soundFilePath, (float)_settings.Alarms.AlarmVolume, "samwise");
             _playback[alarm.Key] = handle;
 
             if (_settings.Alarms.FlashWindow)
             {
                 var win = Application.Current?.MainWindow;
-                if (win is not null) FlashWindow(win);
+                if (win is not null) WindowFlasher.Flash(win);
             }
 
             AlarmTriggered?.Invoke(this, alarm);
@@ -129,33 +130,5 @@ public sealed partial class AlarmService : IDisposable
     {
         foreach (var h in _playback.Values) h.Stop();
         _playback.Clear();
-    }
-
-    [StructLayout(LayoutKind.Sequential)]
-    private struct FLASHWINFO
-    {
-        public uint cbSize;
-        public IntPtr hwnd;
-        public uint dwFlags;
-        public uint uCount;
-        public uint dwTimeout;
-    }
-
-    [LibraryImport("user32.dll")]
-    [return: MarshalAs(UnmanagedType.Bool)]
-    private static partial bool FlashWindowEx(ref FLASHWINFO pwfi);
-
-    private static void FlashWindow(Window window)
-    {
-        var helper = new System.Windows.Interop.WindowInteropHelper(window);
-        var fi = new FLASHWINFO
-        {
-            cbSize = (uint)Marshal.SizeOf<FLASHWINFO>(),
-            hwnd = helper.Handle,
-            dwFlags = 0x0000000F, // FLASHW_ALL | FLASHW_TIMERNOFG = 3 | 12
-            uCount = 5,
-            dwTimeout = 0,
-        };
-        FlashWindowEx(ref fi);
     }
 }
